requirements.txt
# requirements.txt for DefectDojo using Python 3.x
asteval==0.9.18
bleach>=3.1.0
celery==4.4.6
coverage==5.2
defusedxml==0.6.0
django_celery_results==1.2.1
django-auditlog==0.4.7
git+https://github.com/DefectDojo/django-custom-field.git@4e095dd9119858e90fd21560e5618c6347e4ef1b#egg=django-custom-field
django-dbbackup>=3.2.0
django-environ==0.4.5
django-filter==2.3.0
django-imagekit==4.0.2
django-multiselectfield==0.1.12
django-polymorphic==2.1.2
django-crispy-forms==1.9.2
django_extensions==3.0.3
django-rest-swagger==2.2.0
django-slack==5.14.4
django-tagging==0.5.0
git+https://github.com/DefectDojo/django-tastypie-swagger.git@c7c5629480dc20cadd3015490a01bd8e70393c36#egg=django-tastypie-swagger
django-tastypie==0.14.2
django-watson==1.5.5
django-prometheus==2.0.0
Django==2.2.14
djangorestframework==3.11.0
gunicorn==20.0.4
html2text==2020.1.16
humanize==2.5.0
jira==2.0.0
PyGithub==1.51
lxml==4.5.2
Markdown==3.2.2
mysqlclient==2.0.1
pandas>=0.25
xlrd>=1.0.0
pdfkit==0.6.1
Pillow==7.2.0  # required by django-imagekit
psycopg2-binary==2.8.5
cryptography==2.9.2
python-dateutil==2.8.1
python-nmap==0.6.1
pytz==2020.1
redis==3.5.3
requests>=2.22.0
sqlalchemy  # Required by Celery broker transport
supervisor==4.2.0
urllib3==1.25.9
vobject==0.9.6.1
whitenoise>=4.1.2
titlecase==1.1.1
social-auth-app-django==4.0.0
Python-jose==3.1.0
gitpython>=2.1.13
ptvsd>=4.2.7
google-auth==1.19.2
google-api-python-client==1.10.0
google-auth-oauthlib==0.4.1
drf_yasg==1.17.1
jsonlines==1.2.0  # requred by yarn audit parser
django-saml2-auth==2.2.1
